// src/common/apis/octopus/sdk.ts
export type Options<
  TData extends TDataShape = TDataShape,
  ThrowOnError extends boolean = boolean,
> = Options2<TData, ThrowOnError> & {
  /**
   * You can provide a client instance returned by `createClient()` instead of
   * individual options. This might be also useful if you want to implement a
   * custom client.
   */
  client?: Client;
  /**
   * You can pass arbitrary values through the `meta` object. This can be
   * used to access values that aren't defined as part of the SDK function.
   */
  meta?: Record<string, unknown>;
};

class HeyApiClient {
  protected client: Client;

  constructor(args: { client: Client }) {
    this.client = args.client;
  }
}

class HeyApiRegistry<T> {
  private readonly defaultKey = "default";

  private readonly instances: Map<string, T> = new Map();

  get(key?: string): T {
    const instance = this.instances.get(key ?? this.defaultKey);
    if (!instance) {
      throw new Error(
        `No SDK client found. Create one with "new Sdk()" to fix this error.`,
      );
    }
    return instance;
  }

  set(value: T, key?: string): void {
    this.instances.set(key ?? this.defaultKey, value);
  }
}

export class Sdk extends HeyApiClient {
  public static readonly __registry = new HeyApiRegistry<Sdk>();

  constructor(args: { client: Client; key?: string }) {
    super(args);
    Sdk.__registry.set(this, args.key);
  }

  /**
   * Check availability
   *
   * Check the availability of the stream.
   */
  public checkCheck<ThrowOnError extends boolean = false>(
    options?: Options<CheckCheckRequest, ThrowOnError>,
  ) {
    return (options?.client ?? this.client).get<
      CheckCheckResponses,
      unknown,
      ThrowOnError
    >({
      requestValidator: async (data) =>
        await CheckCheckRequestSchema.parseAsync(data),
      responseValidator: async (data) =>
        await CheckCheckResponseSchema.parseAsync(data),
      url: "/check",
      ...options,
    });
  }

  /**
   * Ping
   *
   * Ping.
   */
  public pingPing<ThrowOnError extends boolean = false>(
    options?: Options<PingPingRequest, ThrowOnError>,
  ) {
    return (options?.client ?? this.client).get<
      PingPingResponses,
      unknown,
      ThrowOnError
    >({
      requestValidator: async (data) =>
        await PingPingRequestSchema.parseAsync(data),
      responseValidator: async (data) =>
        await PingPingResponseSchema.parseAsync(data),
      url: "/ping",
      ...options,
    });
  }

  /**
   * Ping headers
   *
   * Ping headers.
   */
  public pingHeadping<ThrowOnError extends boolean = false>(
    options?: Options<PingHeadpingRequest, ThrowOnError>,
  ) {
    return (options?.client ?? this.client).head<
      PingHeadpingResponses,
      unknown,
      ThrowOnError
    >({
      requestValidator: async (data) =>
        await PingHeadpingRequestSchema.parseAsync(data),
      responseValidator: async (data) =>
        await PingHeadpingResponseSchema.parseAsync(data),
      url: "/ping",
      ...options,
    });
  }

  /**
   * Reserve a stream
   *
   * Reserve a stream.
   */
  public reserveReserve<ThrowOnError extends boolean = false>(
    options: Options<ReserveReserveRequest, ThrowOnError>,
  ) {
    return (options.client ?? this.client).post<
      ReserveReserveResponses,
      ReserveReserveErrors,
      ThrowOnError
    >({
      requestValidator: async (data) =>
        await ReserveReserveRequestSchema.parseAsync(data),
      responseValidator: async (data) =>
        await ReserveReserveResponseSchema.parseAsync(data),
      url: "/reserve",
      ...options,
      headers: {
        "Content-Type": "application/json",
        ...options.headers,
      },
    });
  }

  /**
   * Get SSE stream
   *
   * Get a stream of Server-Sent Events.
   */
  public sseSubscribe<ThrowOnError extends boolean = false>(
    options?: Options<SseSubscribeRequest, ThrowOnError>,
  ) {
    return (options?.client ?? this.client).sse.get<
      SseSubscribeResponses,
      SseSubscribeErrors,
      ThrowOnError
    >({
      requestValidator: async (data) =>
        await SseSubscribeRequestSchema.parseAsync(data),
      responseValidator: async (data) =>
        await SseSubscribeResponseSchema.parseAsync(data),
      url: "/sse",
      ...options,
    });
  }

  /**
   * Test
   *
   * Test.
   */
  public testTest<ThrowOnError extends boolean = false>(
    options?: Options<TestTestRequest, ThrowOnError>,
  ) {
    return (options?.client ?? this.client).get<
      TestTestResponses,
      TestTestErrors,
      ThrowOnError
    >({
      requestValidator: async (data) =>
        await TestTestRequestSchema.parseAsync(data),
      responseValidator: async (data) =>
        await TestTestResponseSchema.parseAsync(data),
      url: "/test",
      ...options,
    });
  }
}