// src/common/apis/whip/sdk.ts
// This file is auto-generated by @hey-api/openapi-ts

import type { Client, Options as Options2, TDataShape } from "./client";
import {
  CreateSessionRequestSchema,
  CreateSessionResponseSchema,
  DeleteSessionRequestSchema,
} from "./schemas";
import type {
  CreateSessionErrors,
  CreateSessionRequest,
  CreateSessionResponses,
  DeleteSessionErrors,
  DeleteSessionRequest,
  DeleteSessionResponses,
} from "./types";

export type Options<
  TData extends TDataShape = TDataShape,
  ThrowOnError extends boolean = boolean,
> = Options2<TData, ThrowOnError> & {
  /**
   * You can provide a client instance returned by `createClient()` instead of
   * individual options. This might be also useful if you want to implement a
   * custom client.
   */
  client?: Client;
  /**
   * You can pass arbitrary values through the `meta` object. This can be
   * used to access values that aren't defined as part of the SDK function.
   */
  meta?: Record<string, unknown>;
};

class HeyApiClient {
  protected client: Client;

  constructor(args: { client: Client }) {
    this.client = args.client;
  }
}

class HeyApiRegistry<T> {
  private readonly defaultKey = "default";

  private readonly instances: Map<string, T> = new Map();

  get(key?: string): T {
    const instance = this.instances.get(key ?? this.defaultKey);
    if (!instance) {
      throw new Error(
        `No SDK client found. Create one with "new Sdk()" to fix this error.`,
      );
    }
    return instance;
  }

  set(value: T, key?: string): void {
    this.instances.set(key ?? this.defaultKey, value);
  }
}

export class Sdk extends HeyApiClient {
  public static readonly __registry = new HeyApiRegistry<Sdk>();

  constructor(args: { client: Client; key?: string }) {
    super(args);
    Sdk.__registry.set(this, args.key);
  }

  /**
   * Create a session
   *
   * Create a session to ingest a stream
   */
  public createSession<ThrowOnError extends boolean = false>(
    options: Options<CreateSessionRequest, ThrowOnError>,
  ) {
    return (options.client ?? this.client).post<
      CreateSessionResponses,
      CreateSessionErrors,
      ThrowOnError
    >({
      bodySerializer: null,
      parseAs: "text",
      requestValidator: async (data) =>
        await CreateSessionRequestSchema.parseAsync(data),
      responseValidator: async (data) =>
        await CreateSessionResponseSchema.parseAsync(data),
      url: "/whip/endpoint",
      ...options,
      headers: {
        "Content-Type": "application/sdp",
        ...options.headers,
      },
    });
  }

  /**
   * Delete a session
   *
   * Delete a session and stop the stream
   */
  public deleteSession<ThrowOnError extends boolean = false>(
    options: Options<DeleteSessionRequest, ThrowOnError>,
  ) {
    return (options.client ?? this.client).delete<
      DeleteSessionResponses,
      DeleteSessionErrors,
      ThrowOnError
    >({
      requestValidator: async (data) =>
        await DeleteSessionRequestSchema.parseAsync(data),
      url: "/whip/resource/{session}",
      ...options,
    });
  }
}